TCBZ3488: Inspect memory guarded with pool headers
This change will inspect memory guarded with pool headers. This can avoid errors when a pool content happens to be on a page boundary.

i.e. 0 sized pool could cause an allocated buffer directly points into a guard page, which is legal. However, trying to free this pool will cause the routine changed in this commit to read XP pages, which leads to page fault.