Merged PR 5486: BaseTools/build.py: Fix error propagation on command return code
## Description

Commit ca72f1e introduced a change to output the respfile.txt
contents if the file is detected in a module's OUTPUT directory.

This file is used to store the command string when the length
exceeds 4096 characters so printing the contents of this file helps
place information like the command arguments (build flags) into the
log output regardless of the command length.

However, the change also placed the conditional check to log an error
based on the process return code within the condition that a
respfilelist.txt file exists.

This caused the error message for the return code to not be logged
which, in turn, prevented an exception from being raised which breaks
the build process due to the error.

This change moves the check for the process return code out of that
conditional block so it is always checked regardless of whether a
respfilelist.txt exists.

- [x] Breaking change?
  - Will this change break pre-existing builds or functionality without action being taken?
  **POSSIBLY** - See Integration Instructions

## How This Was Tested

Verify that:
1. Commands from both the compiler and linker show as failed in the build
    log but do not cause the build to fail. The CompilerPlugin reports success.
2. The underlying process return code in failing cases does contain a non-zero
    return code.
3. Reverting commit ca72f1e resolves the problem.
4. Not all modules OUTPUT directories have a respfilelist.txt file present.
5. This change causes the build failure to be recognized and stop the build
    process as expected.

## Integration Instructions

Failures might have been missed previously due to the issue this change
fixes. This is more likely in repositories that built code that did not go into
a flash file (FDF). In any case, uncaught build failures might need to be
fixed when integrating this change.