Merged PR 5625: Policy autogen tool's compatibility on Linux
## Description

This change fixed two path separator caused incompatibility issues when the plugin script is used on Linux systems.

The inline specifier is also added to autogen functions in order to suppress the GCC warnings.

- [ ] Breaking change?
  - Will this change break pre-existing builds or functionality without action being taken?

## How This Was Tested

The change was validated on mu_tiano_platform for both GCC and VS build process.

## Integration Instructions

N/A