Merged PR 5674: Override Processing of directories was failing. Updated processing to account...
## Description

Override validation of directories was failing.
Inside of an INF file, an override string was provided that pointed to a directory, then the override validation would always fail during build.

OverrideValidation.py -w . -t <path to directory that needs to be hash>

This change checks if the override string is a directory before it attempts to process the override string as a file.
Also updated some error strings for better read ability of error messages.

This should not be a breaking change.

## How This Was Tested

-Including override string to a directory in an INF file, build success
-Modifying the hash in the INF file and ensuring a build fail with override message
- Restoring override string to directory, and modifying file in base of folder, ensuring build fail with override message.

## Integration Instructions
Replace file with updated version, and test in a project that contains overrides. Or create overrides.