REBASE: Fix call to clear access attributes when image protection fails
## Description

Images which fail to be protected need their access attributes reset in case access protections were applied to the pages on allocation. The call to ClearAccessAttributesFromMemoryRange() should use LoadedImage, not ImageRecord, to get the base and size parameters. Also, the size parameter should be page aligned before calling ClearAccessAttributesFromMemoryRange().

## How This Was Tested

Booting to shell

## Integration Instructions

N/A