Merged PR 5739: WinRcPath raises exception in failure case
## Description
WinRcPath plugin threw an exception in the failing case. REference to un

Change Description:
When running WinRcPath plugin on a system that did not have rc available, the exception was raised 'PlatformBuilder" has no attribute 'logging'.

thebuilder.logging was probably meant to reference thebuilder.Logger.  Rather than rely on another tool to ensure the logging was available, this change will follow the other plugins and use logging.

Additionaly, the plugin was always returning a success condition, even when RC was not found.  The raised exception would cause the system build to fail when it was not found. The build will still fail when RC is not found by returning a failure value.

[ ] Breaking change?
This will not be a breaking change

## How This Was Tested

Went through all the Windows kits installed on the system, and renamed rc.exe to rc.exe.old so the vswhere would fail
Windows Kits\10\bin\10.0.17763.0\x86
Windows Kits\10\bin\10.0.17763.0\x64
Windows Kits\10\bin\10.0.17763.0\x86
Windows Kits\10\bin\10.0.17763.0\x64
Windows Kits\10\bin\10.0.19041.0\x86
Windows Kits\10\bin\10.0.19041.0\x64

Once the VsWhere failed, ran the Q35 build to ensure that the plugin reported a failure when the tool was not found.

Tested Win 11 VS2019 and above windows kits.

## Integration Instructions

Replace existing file with updated version.