Update Extra Buffer Case When Separating Images in the Memory Map
Description
Once we no longer have found an image contained within a buffer, the end of the previous loop will already have set PhysicalStart to be the end of the last image found. Therefore, the remaining memory range (PhysicalEnd - PhysicalStart) should not be any part of a loaded image memory range and so it should not be designated as a code/data region via attributes.

Breaking Change?
No

How This Was Tested
Booting to shell

Integration Instructions
N/A