BUG: NULL Detection (Page 0 RP) Not Working -- Update EnableNullDetection() to TPL_CALLBACK - 1
## Description

Previous commits updated InitializeDxeNxMemoryProtectionPolicyMu() to no longer fill out the gCpu global if it was NULL. Instead, the InitializeDxeNxMemoryProtectionPolicyMu() function called on Cpu Arch Protocol installation was updated to TPL_CALLBACK - 1 so the DxeCore GenericProtocolNotify() logic could populate gCpu. A separate event initialized NULL detection and was still set to TPL_CALLBACK and, because InitializeDxeNxMemoryProtectionPolicyMu() no longer filled out the global, the NULL page attributes could not be updated when EnableNullDetection() was called.

This PR moves the event which calls EnableNullDetection() on Cpu Arch Protocol installation to TPL_CALLBACK - 1.

## Breaking change?
N/A

## How This Was Tested

Booting to shell and running the memory protection test app

## Integration Instructions

N/A