Move Attribute Clear Outside of Memory Lock
## Description

It's possible that clearing the access attributes from a memory range requires the allocation of page table memory. The page table memory allocation will call to AllocatePages() which will then try to acquire the memory lock and ASSERT. This PR moves the access attribute clear outside of the lock acquisition to fix this case.

## Breaking change?
No

## How This Was Tested

Running on Arcata

## Integration Instructions

N/A